+++ lib/list_operators.dart
@@ -1,11 +1,12 @@
 export 'src/exceptions/empty_iterable.dart';
 export 'src/exceptions/length_mismatch.dart';
-export 'src/extensions/must_have.dart';
 export 'src/extensions/comparison.dart';
 export 'src/extensions/double_operators.dart';
 export 'src/extensions/euclidean_metric.dart';
 export 'src/extensions/export.dart';
 export 'src/extensions/int_operators.dart';
+export 'src/extensions/must_have.dart';
 export 'src/extensions/num_operators.dart';
+export 'src/extensions/spherical_coordinates.dart';
 export 'src/extensions/unmodifiable.dart';
 export 'src/matchers/close_to_list.dart';